requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzdata==2024.1
yarl==1.9.4
